refactor : 예약이 있을시 Theme 삭제 불가능하도록 예외처리

+++ src/main/java/nextstep/roomescape/theme/service/ThemeServiceImpl.java
@@ -1,6 +1,9 @@
 package nextstep.roomescape.theme.service;
 
 import nextstep.roomescape.exception.NotExistEntityException;
+import nextstep.roomescape.exception.UsedExistEntityException;
+import nextstep.roomescape.reservation.repository.ReservationRepository;
+import nextstep.roomescape.reservation.repository.ReservationRepositoryJdbcImpl;
 import nextstep.roomescape.theme.controller.dto.ThemeRequestDTO;
 import nextstep.roomescape.theme.repository.ThemeRepository;
 import nextstep.roomescape.theme.repository.model.Theme;
@@ -13,11 +16,12 @@ import java.util.List;
 public class ThemeServiceImpl implements ThemeService {
 
     private final ThemeRepository themeRepository;
+    private final ReservationRepository reservationRepository;
     @Autowired
-    public ThemeServiceImpl(ThemeRepository themeRepository) {
+    public ThemeServiceImpl(ThemeRepository themeRepository, ReservationRepositoryJdbcImpl reservationRepository) {
         this.themeRepository = themeRepository;
+        this.reservationRepository = reservationRepository;
     }
-
     @Override
     public Long create(ThemeRequestDTO themeRequestDTO) {
         return themeRepository.create(themeRequestDTO.toEntity());
@@ -33,6 +37,9 @@ public class ThemeServiceImpl implements ThemeService {
         Theme theme = themeRepository.findById(id);
         if (theme == null){
             throw new NotExistEntityException("해당 id 테마는 존재하지 않습니다.");
+        }
+        if (reservationRepository.findByThemeId(id).size() > 0){
+            throw new UsedExistEntityException("해당 테마는 예약이 있어 삭제가 불가능합니다.");
         }
         themeRepository.delete(id);
     }
